conflit resolution final ?

+++ Shoot_em_up/Player.cpp
@@ -60,7 +60,7 @@ void PlayerShip::render(sf::RenderWindow& window)
 
     m_sprite.setRotation(m_angle * 180 / 3.14);
     sf::Vector2u textureSize = m_texture.getSize();
-    m_sprite.setOrigin(textureSize.x / 2, textureSize.y / 2);
+    m_sprite.setOrigin(static_cast<float> (textureSize.x / 2), static_cast<float> (textureSize.y / 2));
     m_sprite.setPosition(m_position.x,m_position.y);
     window.draw(m_sprite);
